refactor beam components into a separate package
