+++ compiler/frontend-beam/src/main/java/edu/snu/onyx/compiler/frontend/beam/BeamKeyExtractor.java
@@ -13,12 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.onyx.client.beam;
+package edu.snu.onyx.compiler.frontend.beam;
 
-import org.apache.beam.sdk.options.PipelineOptions;
+import edu.snu.onyx.common.KeyExtractor;
+import org.apache.beam.sdk.values.KV;
 
 /**
- * OnyxPipelineOptions.
+ * Extracts the key from a KV element.
+ * For non-KV elements, the elements themselves become the key.
  */
-public interface OnyxPipelineOptions extends PipelineOptions {
+final class BeamKeyExtractor implements KeyExtractor {
+  /**
+   * Extracts key.
+   * @param element element to get the key from.
+   * @return the extracted key of the element.
+   */
+  public Object extractKey(final Object element) {
+    if (element instanceof KV) {
+      return ((KV) element).getKey();
+    } else {
+      return element;
+    }
+  }
 }
