+++ compiler/frontend-beam/src/main/java/edu/snu/onyx/compiler/frontend/beam/OnyxPipelineOptions.java
@@ -13,26 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.onyx.client.beam;
+package edu.snu.onyx.compiler.frontend.beam;
 
-import edu.snu.onyx.common.KeyExtractor;
-import org.apache.beam.sdk.values.KV;
+import org.apache.beam.sdk.options.PipelineOptions;
 
 /**
- * Extracts the key from a KV element.
- * For non-KV elements, the elements themselves become the key.
+ * OnyxPipelineOptions.
  */
-final class BeamKeyExtractor implements KeyExtractor {
-  /**
-   * Extracts key.
-   * @param element element to get the key from.
-   * @return the extracted key of the element.
-   */
-  public Object extractKey(final Object element) {
-    if (element instanceof KV) {
-      return ((KV) element).getKey();
-    } else {
-      return element;
-    }
-  }
+public interface OnyxPipelineOptions extends PipelineOptions {
 }
