+++ compiler/frontend-beam/src/main/java/edu/snu/onyx/compiler/frontend/beam/OnyxPipelineVisitor.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.onyx.client.beam;
+package edu.snu.onyx.compiler.frontend.beam;
 
 import edu.snu.onyx.compiler.frontend.beam.coder.BeamCoder;
 import edu.snu.onyx.common.dag.DAGBuilder;
