+++ compiler/frontend-beam/src/main/java/edu/snu/onyx/compiler/frontend/beam/coder/BeamCoder.java
