+++ compiler/frontend-beam/src/main/java/edu/snu/onyx/compiler/frontend/beam/transform/FlattenTransform.java
