+++ examples/src/main/java/edu/snu/onyx/examples/beam/MapReduce.java
@@ -15,8 +15,8 @@
  */
 package edu.snu.onyx.examples.beam;
 
-import edu.snu.onyx.client.beam.OnyxPipelineOptions;
-import edu.snu.onyx.client.beam.OnyxPipelineRunner;
+import edu.snu.onyx.compiler.frontend.beam.OnyxPipelineOptions;
+import edu.snu.onyx.compiler.frontend.beam.OnyxPipelineRunner;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
 import org.apache.beam.sdk.options.PipelineOptionsFactory;
