+++ examples/src/main/java/edu/snu/onyx/examples/beam/MultinomialLogisticRegression.java
@@ -16,7 +16,7 @@
 package edu.snu.onyx.examples.beam;
 
 import edu.snu.onyx.compiler.frontend.beam.transform.LoopCompositeTransform;
-import edu.snu.onyx.client.beam.OnyxPipelineRunner;
+import edu.snu.onyx.compiler.frontend.beam.OnyxPipelineRunner;
 import edu.snu.onyx.common.Pair;
 import org.apache.beam.sdk.Pipeline;
 import org.apache.beam.sdk.options.PipelineOptions;
